+++ .github/workflows/gcd-upload.yaml
@@ -5,6 +5,7 @@ on:
     branches:
       - workflow-testing
       - main
+      - alerts-and-cleanup
 
 jobs:
   upload_to_gcs:
