+++ data_pipeline/dags/train_data_dag.py
@@ -1,29 +1,25 @@
 from datetime import datetime, timedelta
 from airflow import DAG
 from airflow.operators.python import PythonOperator
-from scripts.data_utils import upload_train_data_to_gcs  # Reusing existing utility
 import logging
 import logging
 from airflow.operators.email import EmailOperator
 from airflow.operators.python import PythonOperator
 from airflow.operators.dagrun_operator import TriggerDagRunOperator
 
-from scripts.bigquery_utils import (
+from scripts.bq.bigquery_utils import (
     check_sample_count,
     get_bq_data,
     perform_similarity_search,
     upload_gcs_to_bq
 )
-from scripts.data_processing import (
+from scripts.data.data_processing import (
     get_initial_queries,
     
 )
 from scripts.llm_utils import generate_llm_response
 
-from scripts.data_utils import (
-    upload_to_gcs,
-    upload_train_data_to_gcs
-)       
+from scripts.gcs.gcs_utils import upload_train_data_to_gcs, upload_to_gcs     
 
 logging.basicConfig(level=logging.INFO)
